+++ src/controllers/GameHistory.js
@@ -73,6 +73,8 @@ exports.isDeathPotionUsed = async gameId => !!await this.findOne({ gameId, "play
 
 exports.isInfectionUsed = async gameId => !!await this.findOne({ gameId, "play.targets.isInfected": true });
 
+exports.isSecondVoteRequestUsed = async gameId => !!await this.findOne({ gameId, "play.doesJudgeRequestAnotherVote": true });
+
 exports.getLastNightPlay = gameId => {
     const nightPlayActions = [...turnPreNightActionsOrder, ...turnNightActionsOrder].map(({ action }) => action);
     return this.findOne({ gameId, "play.action": { $in: nightPlayActions } }, null, { sort: { createdAt: -1 } });